requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
configparser==4.0.2
contextlib2==0.6.0.post1
greenlet==1.1.1
idna==3.2
importlib-metadata==2.1.1
mysqlclient==2.0.3
pathlib2==2.3.6
PyMySQL==1.0.2
pyTelegramBotAPI==3.8.2
python-decouple==3.4
requests==2.26.0
scandir==1.10.0
six==1.16.0
SQLAlchemy==1.4.23
urllib3==1.26.6
zipp==1.2.0
